Refactor EntradaResponseModel and update input handling
- Changed `HorasExtras` from `int` to `decimal` in `EntradaResponseModel` for precision.
- Added `imask` script in `Fixos.cshtml` and applied input masking to `Periodo`.
- Updated input types for `ValorHora` and `HorasExtras` in `RegistrarEntrada.cshtml` to `text` with `inputmode="decimal"`.
- Enhanced `MesAno` input field with an `id` for easier access.
- Modified calculation of `valorExtraCalculado` in `RegistrarEntrada.cshtml.cs` to use `decimal` for `HorasExtras`.
- Changed `HorasExtras` in `InputModel` from `int?` to `decimal?` for consistency.

+++ Portal.ControleFinanceiro/Models/Response/EntradaResponseModel.cs
@@ -4,7 +4,7 @@
     {
         public string Message { get; set; }
         public decimal ValorHoraExtra { get; set; }
-        public int HorasExtras { get; set; }
+        public decimal HorasExtras { get; set; }
         public decimal ValorExtraCalculado { get; set; }
         public decimal NovosExtras { get; set; }
     }
